[IMP] Moved file into app.blueprints.rest namespace
