+++ source/app/blueprints/rest/case/case_assets_routes.py
@@ -24,7 +24,7 @@ from flask import request
 from flask_login import current_user
 
 from app import db
-from app.blueprints.case.case_comments import case_comment_update
+from app.blueprints.rest.case_comments import case_comment_update
 from app.datamgmt.case.case_assets_db import add_comment_to_asset
 from app.datamgmt.case.case_assets_db import create_asset
 from app.datamgmt.case.case_assets_db import delete_asset
