+++ source/app/blueprints/rest/case/case_tasks_routes.py
@@ -24,7 +24,7 @@ from flask import request
 from flask_login import current_user
 
 from app import db
-from app.blueprints.case.case_comments import case_comment_update
+from app.blueprints.rest.case_comments import case_comment_update
 from app.blueprints.rest.endpoints import response_api_deleted
 from app.blueprints.rest.endpoints import endpoint_deprecated
 from app.blueprints.rest.endpoints import response_api_error
