+++ source/app/blueprints/rest/case/case_timeline_routes.py
@@ -29,7 +29,7 @@ from sqlalchemy import and_
 
 from app import db
 from app import app
-from app.blueprints.case.case_comments import case_comment_update
+from app.blueprints.rest.case_comments import case_comment_update
 from app.datamgmt.case.case_assets_db import get_asset_by_name
 from app.datamgmt.case.case_events_db import add_comment_to_event
 from app.datamgmt.case.case_events_db import get_category_by_name
